Handle omitted examples
If the example field is blank, as in {{ lib // // with no _ // example
}}, process it with the example property set to '' in the lib.

This is so that when another component (the API or the UI) receives this lib, it can show no example at all, and not even say "For example,".

Examples may color the player's thinking. It's funnier when people bring
their own concept of what the prompt is about.

madliberation-130